requirements.txt
numpy==1.23.3
opencv_python==4.10.0.84
pyntcore==2024.3.2.1
PyYAML==6.0.2
robotpy_apriltag==2024.3.2.1
robotpy_cscore==2024.3.2.1
robotpy_wpimath==2024.3.2.1
typing_extensions==4.12.2
wpilib==2024.3.2.1
